+++ src/components/queryBuilder/views/LogsQueryBuilder.tsx
@@ -94,7 +94,7 @@ export const LogsQueryBuilder = (props: LogsQueryBuilderProps) => {
   }, builderState);
 
   useLogDefaultsOnMount(datasource, isNewQuery, builderOptions, builderOptionsDispatch);
-  useOtelColumns(builderState.otelEnabled, builderState.otelVersion, builderOptionsDispatch);
+  useOtelColumns(datasource, builderState.otelEnabled, builderState.otelVersion, builderOptionsDispatch);
   useDefaultTimeColumn(datasource, allColumns, builderOptions.table, builderState.timeColumn, builderState.otelEnabled, builderOptionsDispatch);
   useDefaultFilters(builderOptions.table, isNewQuery, builderOptionsDispatch);