+++ src/selectors.ts
@@ -133,6 +133,10 @@ export const Components = {
       aliasTableInput: 'config__alias-table-config__alias-table-input',
     }
   },
+  LogsContextPanel: {
+    alert: 'logs-context-panel__alert',
+    LogsContextKey: 'logs-context-panel__logs-context-key',
+  },
   QueryBuilder: {
     expandBuilderButton: 'query-builder__expand-builder-button',
     LogsQueryBuilder: {