+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features
 
+- Implemented log context for log queries
+- Added configuration options for log context columns
 - Queries parsed from the SQL editor will now attempt to re-map columns into their correct fields for Log and Trace queries.
 
 ### Fixes